move templates to templates/index.js
